+++ operator/internal/controller/aisauth/doc.go
@@ -1,0 +1,6 @@
+/*
+ * Copyright (c) 2026, NVIDIA CORPORATION. All rights reserved.
+ */
+
+// Package aisauth contains Kubernetes controller logic for AIStoreAuth resources.
+package aisauth